// frontend/src/web-backend-api.d.ts
export interface paths {
// ...
    "/api/v1/devices/self": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Get all devices owned by a given user */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["DeviceDTO"][];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/api/v1/pets/self/{petId}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Get a pet by its id */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    petId: number;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["PetDTO"];
                    };
                };
                /** @description Not Found */
                404: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        /**
         * Format: date
         * @example 2022-03-10
         */
        Date: string;
        DeviceDTO: {
            /** Format: int64 */
            id?: number;
            name?: string;
            type?: string;
            ownerId?: string;
            /** Format: int64 */
            petId?: number;
        };
        PetDTO: {
            /** Format: int64 */
            id?: number;
            name?: string;
            /** Format: int64 */
            type?: number;
            ownerId?: string;
            lastTimeOnDevice?: components["schemas"]["Date"];
            state?: components["schemas"]["PetStateDTO"];
        };
        PetStateDTO: {
            /** Format: int32 */
            happiness?: number;
            /** Format: int32 */
            wellbeing?: number;
            /** Format: int32 */
            health?: number;
            /** Format: int32 */
            hunger?: number;
            /** Format: int32 */
            cleanliness?: number;
            /** Format: int32 */
            fun?: number;
            /** Format: int32 */
            xp?: number;
        };
        PetTypeDTO: {
            /** Format: int64 */
            id?: number;
            name?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}